New pipeline script that allows parallel ingestion
In here it's included a script that allows to run the ingestion of the
xml files into a ccdata R file and into an anonymised one.

The files are broken into multiple ones based in the number of patients,
then they are converted into ccdata by doing multiple at a time using
xargs.

New function to xml2ccdata was added to remove suffixes to filenames in
order to keep the original filename in the data header.